filtergw: add plugin allow to allow/block gateways (even smartgateway)
The plugin adds a possibility to filter out gateway HNAs. This can be
very handy e.g. for faulty nodes. A gateway can be taken out directly on
a central node and all traffic is then redirected. Also the underlying
smart gateways are adapted. Currently the link multiplier is often used
to control a faulty node, but it is not necessarily the best way to do
so.

The plugin also allows to determine the smart gateway on the local node.
This way you can easily select another gateway node to direct traffic,
or block a specific one.

Here are some sample configurations:

Only allow Gateway HNA from '10.230.132.40':
config LoadPlugin
        option library 'olsrd_filtergw'
        list originator '10.230.132.40'
        option allowlist '1'

Block Gateway HNA from '10.230.132.40':
config LoadPlugin
        option library 'olsrd_filtergw'
        list originator '10.230.132.40'
        option allowlist '0'

The plugin is based on the work of svenroad, pmelange and booo.

Signed-off-by: Nick Hainke <vincent@systemli.org>